[REFACTOR]: ️ Workspace에서 유저 관련 기존 로직 제거 (#89)
* [REFACTOR]: API 리팩토링

* [TEST]: 미사용 테스트 제거

* [REFACTOR]: 워크스페이스 유저 권한 로직 추가

* [REFACTOR]: 워크스페이스 유저 권한 검증 로직 추가

* [REFACTOR]: 워크스페이스 유저 관련 로직 제거

* [REFACTOR]: UserWorkspace 삭제

* [REFACTOR]: 유저 워크스페이스 삭제

* [TEST]: 테스트 임시 변경

* [REFACTOR]: 테스트 전용 상수로 변경

* [REFACTOR]: 하드코딩된 검증 로직 변경

+++ src/main/java/project/domain/workspaceUser/service/WorkspaceUserService.java
@@ -6,16 +6,18 @@ import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
 import project.common.exception.ErrorCode;
 import project.common.exception.PlantException;
+import project.common.util.UserUtil;
+import project.common.util.WorkspaceUserUtil;
 import project.domain.user.dao.UserRepository;
 import project.domain.workspace.dao.WorkspaceRepository;
 import project.domain.workspace.domain.Workspace;
 import project.domain.workspaceUser.dao.WorkspaceUserRepository;
 import project.domain.workspaceUser.domain.WorkspaceUser;
+import project.domain.workspaceUser.domain.WorkspaceUserDomainService;
 import project.domain.workspaceUser.dto.request.WorkspaceUserCreateRequest;
 import project.domain.workspaceUser.dto.request.WorkspaceUserUpdateRequest;
 import project.domain.workspaceUser.dto.response.WorkspaceUsersResponse;
 
-// TODO: 권한 검증 로직 추가
 @Service
 @RequiredArgsConstructor
 public class WorkspaceUserService {
@@ -23,9 +25,13 @@ public class WorkspaceUserService {
     private final WorkspaceRepository workspaceRepository;
     private final UserRepository userRepository;
     private final WorkspaceUserRepository workspaceUserRepository;
+    private final UserUtil userUtil;
+    private final WorkspaceUserUtil workspaceUserUtil;
+    private final WorkspaceUserDomainService workspaceUserDomainService;
 
     @Transactional(readOnly = true)
     public WorkspaceUsersResponse findWorkspaceUsersByWorkspace(Long workspaceId) {
+        validateLoginUserInWorkspace(workspaceId);
         Workspace workspace = workspaceRepository
                 .findById(workspaceId)
                 .orElseThrow(() -> new PlantException(ErrorCode.WORKSPACE_NOT_FOUND));
@@ -35,6 +41,7 @@ public class WorkspaceUserService {
 
     @Transactional
     public Long addUserToWorkspace(WorkspaceUserCreateRequest request) {
+        validateLoginUserInWorkspace(request.workspaceId());
         var workspace = workspaceRepository
                 .findById(request.workspaceId())
                 .orElseThrow(() -> new PlantException(ErrorCode.WORKSPACE_NOT_FOUND));
@@ -42,7 +49,11 @@ public class WorkspaceUserService {
                 .findById(request.userId())
                 .orElseThrow(() -> new PlantException(ErrorCode.USER_NOT_FOUND));
 
-        var workspaceUser = WorkspaceUser.create(workspace, user);
+        if (workspaceUserRepository.existsByWorkspaceIdAndUserId(workspace.getId(), user.getId())) {
+            throw new PlantException(ErrorCode.WORKSPACE_USER_ALREADY_EXIST);
+        }
+
+        var workspaceUser = WorkspaceUser.createUser(workspace, user);
         workspaceUserRepository.save(workspaceUser);
         return workspaceUser.getId();
     }
@@ -52,7 +63,13 @@ public class WorkspaceUserService {
         WorkspaceUser workspaceUser = workspaceUserRepository
                 .findById(workspaceUserId)
                 .orElseThrow(() -> new PlantException(ErrorCode.WORKSPACE_USER_NOT_FOUND));
+        validateLoginUserIsAdmin(workspaceUser.getWorkspace().getId());
+
+        List<WorkspaceUser> workspaceUsers = workspaceUserRepository.findAllByWorkspaceId(
+                workspaceUser.getWorkspace().getId());
+        workspaceUserDomainService.validateWhenChangeRole(workspaceUser, request.role(), workspaceUsers);
         workspaceUser.updateRole(request.role());
+
         workspaceUserRepository.save(workspaceUser);
     }
 
@@ -61,7 +78,29 @@ public class WorkspaceUserService {
         WorkspaceUser workspaceUser = workspaceUserRepository
                 .findById(workspaceUserId)
                 .orElseThrow(() -> new PlantException(ErrorCode.WORKSPACE_USER_NOT_FOUND));
+        List<WorkspaceUser> workspaceUsers = workspaceUserRepository.findAllByWorkspaceId(
+                workspaceUser.getWorkspace().getId());
 
+        validateLoginUserIsAdmin(workspaceUser.getWorkspace().getId());
+        workspaceUserDomainService.validateWhenRemoveUser(workspaceUser, workspaceUsers);
         workspaceUserRepository.delete(workspaceUser);
+    }
+
+    private void validateLoginUserIsAdmin(Long workspaceId) {
+        Long loginUserId = userUtil.getLoginUserId();
+        boolean isAdmin = workspaceUserUtil.isAdminUser(workspaceId, loginUserId);
+
+        if (!isAdmin) {
+            throw new PlantException(ErrorCode.WORKSPACE_USER_AUTHORITY_INVALID);
+        }
+    }
+
+    private void validateLoginUserInWorkspace(Long workspaceId) {
+        Long loginUserId = userUtil.getLoginUserId();
+        boolean exists = workspaceUserUtil.existsUser(workspaceId, loginUserId);
+
+        if (!exists) {
+            throw new PlantException(ErrorCode.WORKSPACE_USER_AUTHORITY_INVALID);
+        }
     }
 }
